requirements.txt
python-jose[cryptography]
passlib[bcrypt]
python-multipart
fastapi[all] --only-binary :all:
bcrypt
